+++ src/core/contract/ServerContract.cpp
@@ -49,7 +49,6 @@
 #include "opentxs/core/util/Assert.hpp"
 
 #include <stdint.h>
-#include <zcert.h>
 #include <list>
 #include <memory>
 #include <string>
